+++ Backend/src/LoanLedger.Application/Authentication/AuthService.cs
@@ -8,14 +8,17 @@ public class AuthService : IAuthService
 {
     private readonly IUserRepository _repository;
     private readonly IPasswordService _passwordService;
+	private readonly IJwtTokenService _jwtTokenService;
 
     public AuthService(
-        IUserRepository repository,
-        IPasswordService passwordService)
-    {
-        _repository = repository;
-        _passwordService = passwordService;
-    }
+		IUserRepository repository,
+		IPasswordService passwordService,
+		IJwtTokenService jwtTokenService)
+	{
+		_repository = repository;
+		_passwordService = passwordService;
+		_jwtTokenService = jwtTokenService;
+	}
 
     public async Task<RegisterUserResponse> RegisterAsync(RegisterUserRequest request)
 	{
@@ -75,6 +78,56 @@ public class AuthService : IAuthService
 	}
 	public async Task<LoginResponse> LoginAsync(LoginRequest request)
 	{
-		throw new NotImplementedException();
+		var user = await _repository.GetByEmailOrPhoneAsync(request.EmailOrPhone);
+
+		if (user == null)
+		{
+			return new LoginResponse
+			{
+				Success = false,
+				Message = "Invalid email/phone or password."
+			};
+		}
+
+		if (!user.IsActive)
+		{
+			return new LoginResponse
+			{
+				Success = false,
+				Message = "Account has been disabled."
+			};
+		}
+
+		var validPassword =
+			_passwordService.VerifyPassword(
+				user.PasswordHash,
+				request.Password);
+
+		if (!validPassword)
+		{
+			return new LoginResponse
+			{
+				Success = false,
+				Message = "Invalid email/phone or password."
+			};
+		}
+
+		var token =
+			_jwtTokenService.GenerateToken(
+				user.Id,
+				user.Email);
+
+		user.LastLoginAt = DateTime.UtcNow;
+
+		await _repository.SaveChangesAsync();
+
+		return new LoginResponse
+		{
+			Success = true,
+			Message = "Login successful.",
+			Token = token,
+			UserId = user.Id,
+			FullName = user.FullName
+		};
 	}
 }